requirements.txt
requests
phonenumbers
mcstatus
python-nmap
colorama
beautifulsoup4
